Don't do pokemon exception handling
"Gotta catch 'em all" does not apply to exceptions. Just got a crash here, and
the output is not what we'd want to get as part of an issue, or helpful to the
user in any way:

    The following patterns will be excluded: ['.DS_Store', 'Thumbs.db']
    string indices must be integers

Instead, add a new exception base class for errors which are not bugs, and only
catch those.

This also stops everything from inheriting ConnectException - I don't see why a
Print- or ConfigurationException should be a ConnectException.